requirements.txt
opencv-python
pandas
numpy
pyqt5
sklearn
qimage2ndarray
pillow
matplotlib
tensorflow